Fix to_matrix_blocked dtype: float32 (kornia requires Float)
The previous commit changed to_matrix_blocked to allocate the 2x3
matrix in float64. Original code used torch.empty((2, 3)) which
defaults to float32, and kornia.warp_affine on the SW path errors
out with "expected scalar type Float but found Double". Restore
float32 — same numerical layout as before.

+++ ETNA/registrators_pyramidal.py
@@ -323,11 +323,13 @@ class EtnaMultiMetric(object):
             theta = float(vector_params[2])
         cs = math.cos(theta)
         sn = math.sin(theta)
-        # Single allocation in float64 — was 11 separate torch ops on 0-d
-        # tensors (~5 ms each on ARM CPU). Bit-equivalent.
+        # Single allocation. Default dtype is float32 — matches the original
+        # torch.empty((2, 3)) layout (kornia.warp_affine requires float32),
+        # while the input scalars come from a float64 par_lin so the values
+        # round-trip the same way they did before.
         mat_params = torch.tensor(
             [[cs, sn, tx], [-sn, cs, ty]],
-            dtype=torch.float64, device=self.device,
+            dtype=torch.float32, device=self.device,
         )
         self._wrap_times["to_matrix_blocked"] += time.perf_counter() - _t
         self._wrap_counts["to_matrix_blocked"] += 1
